Extract builder class to a separated file
